fix: delete constant on app
