+++ apps/platform/src/pages/CredibleSetPage/ProfileHeader.tsx
@@ -16,7 +16,7 @@ import {
 import { Box, Typography } from "@mui/material";
 import CREDIBLE_SET_PROFILE_HEADER_FRAGMENT from "./ProfileHeader.gql";
 import { epmcUrl } from "@ot/utils";
-import { credsetConfidenceMap, populationMap } from "../../constants";
+import { credsetConfidenceMap, populationMap } from "@ot/constants";
 
 function ProfileHeader() {
   const { loading, error, data } = usePlatformApi();
