+++ apps/platform/src/pages/HomePage/Splash.jsx
@@ -3,7 +3,7 @@ import { makeStyles } from "@mui/styles";
 import Particles from "react-tsparticles";
 import { loadFull } from "tsparticles";
 
-import { particlesConfig } from "../../constants";
+import { particlesConfig } from "@ot/constants";
 
 const useStyles = makeStyles(theme => ({
   splashContainer: {
